LayoutTests/webrtc/vp9-expected.txt

  

PASS VP9 in getCapabilities 
PASS Verify VP9 activation 
PASS Setting video exchange 
PASS Ensuring connection state is connected 
PASS Track is enabled, video should not be black 
PASS Track is disabled, video should be black 
PASS Track is enabled, video should not be black 2 
